fix: address review feedback for Schema v5: drop vms.health and HealthState
